📜  Switching to Typescript
Adding typescript 3.4.5 and typos
