+++ jest.config.js
@@ -2,28 +2,17 @@ const webConfig = {
     name: "web",
     displayName: "Web Client",
     clearMocks: true,
-    globals: {
+    globals: {},
+    transform: {
+        "^.+\\.ts?$": "ts-jest"
     },
-    moduleDirectories: [
-        "node_modules"
-    ],
-    moduleFileExtensions: [
-        "js",
-        "jsx"
-    ],
-    roots: [
-        ""
-    ],
-    snapshotSerializers: [
-        "enzyme-to-json/serializer"
-    ],
-    testMatch: [
-        "<rootDir>/tests/*.spec.js"
-    ],
-    testPathIgnorePatterns: [
-        "<rootDir>/node_modules"
-    ],
-    reporters: [ "default" ],
+    moduleDirectories: ["node_modules"],
+    moduleFileExtensions: ["ts", "js"],
+    roots: [""],
+    snapshotSerializers: ["enzyme-to-json/serializer"],
+    testMatch: ["<rootDir>/tests/*.spec.[t|j]s"],
+    testPathIgnorePatterns: ["<rootDir>/node_modules"],
+    reporters: ["default"],
     coverageDirectory: "<rootDir>/dist/coverage"
 };
