+++ typings/map-stream.d.ts
@@ -1,0 +1,5 @@
+declare module "map-stream" {
+    function map(dataFunction: (data: any, callback: () => void) => void): string;
+
+    export = map;
+}
